third_party/blink/web_tests/fast/dom/inert/inert-node-is-uneditable.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<span inert id="not-editable" contenteditable>I'm not editable.</span>
<span id="editable" contenteditable>I'm editable.</span>
<script>
var notEditable = document.querySelector('#not-editable');
var editable = document.querySelector('#editable');

promise_test(async () => {
    notEditable.focus();
    var oldValue = notEditable.textContent;
    assert_equals(oldValue, "I'm not editable.");
    eventSender.keyDown('a');
    assert_equals(notEditable.textContent, oldValue);
}, "Can't edit inert contenteditable");

promise_test(async () => {
    editable.focus();
    var oldValue = editable.textContent;
    assert_equals(oldValue, "I'm editable.");
    eventSender.keyDown('a');
    assert_not_equals(editable.textContent, oldValue);
}, "Can edit non-inert contenteditable");

</script>
</body>
</html>
